Move stateless_reset into its own test unit
It has nothing to do with 0rtt, and doesn't make sense to bundle them
together.
